+++ CHANGELOG.md
@@ -9,14 +9,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.15.1] - Unreleased
 
+### Added:
+
+- Diagnostics: Ensure at least one plugin is enabled
+
 ### Changed:
 
 - Snippets: `devproxy-plugin-open-api-spec-generator` - OpenApiSpecGeneratorPlugin config section
 - Snippets: All snippets that reference schemas updated to use `v0.24.0` schema
+- Diagnostics: Refactored code to be more readable and maintainable
+- Diagnostics: Improved config section check
 
 ### Fixed:
 
 - Snippets: Fix invalid Json in `devproxy-config-file`
+- Diagnostics: Fixed update schema check in proxy files
 
 ## [0.14.0] - 2024-11-27